A lot of changes
A series of Initializer objects in progress for when the program initially starts up.
All Finder classes now implement the same interface and can now return both an object or its index.
Started work on menu handling.

This entire project is a mess and I don't have a to-do list for any of this, hooray.